-loading，stop loading, loading fail 初步测试

+++ loading/src/main/java/com/foundation/widget/loading/Extensions.kt
@@ -13,7 +13,7 @@ import android.view.ViewGroup
 internal const val MATCH_PART = ViewGroup.LayoutParams.MATCH_PARENT
 internal const val WRAP_CONTENT = ViewGroup.LayoutParams.WRAP_CONTENT
 internal val Float.dp get() = this * Resources.getSystem().displayMetrics.density+0.5F
-internal val Int.dp get() = this.toFloat().dp
+internal val Int.dp get() = this.toFloat().dp.toInt()
 
 internal fun Int.toExactlyMeasureSpec():Int{
     return View.MeasureSpec.makeMeasureSpec(this, View.MeasureSpec.EXACTLY)
